comp7.1 - oopps, needed to make whole new abc
